+++ src/main/java/org/openlmis/auth/i18n/MessageKeys.java
@@ -49,6 +49,11 @@ public abstract class MessageKeys {
   public static final String PASSWORD_RESET_EMAIL_BODY = SERVICE_PREFIX
       + ".email.resetPassword.body";
 
+  public static final String EMAIL_VERIFICATION_EMAIL_SUBJECT = SERVICE_PREFIX
+      + ".email.verifyEmail.subject";
+  public static final String EMAIL_VERIFICATION_EMAIL_BODY = SERVICE_PREFIX
+      + ".email.verifyEmail.body";
+
   public static final String USERS_PASSWORD_RESET_USER_NOT_FOUND = USERS
       + ".passwordReset.userNotFound";