+++ src/main/java/org/openlmis/auth/service/UserService.java
@@ -17,15 +17,23 @@ package org.openlmis.auth.service;
 
 
 import static org.openlmis.auth.i18n.MessageKeys.ACCOUNT_CREATED_EMAIL_SUBJECT;
+import static org.openlmis.auth.i18n.MessageKeys.EMAIL_VERIFICATION_EMAIL_BODY;
+import static org.openlmis.auth.i18n.MessageKeys.EMAIL_VERIFICATION_EMAIL_SUBJECT;
 import static org.openlmis.auth.i18n.MessageKeys.ERROR_REFERENCE_DATA_USER_NOT_FOUND;
 import static org.openlmis.auth.i18n.MessageKeys.PASSWORD_RESET_EMAIL_BODY;
 import static org.openlmis.auth.i18n.MessageKeys.PASSWORD_RESET_EMAIL_SUBJECT;
 
+import java.time.ZonedDateTime;
+import java.util.function.Function;
+import org.openlmis.auth.domain.EmailVerificationToken;
+import org.openlmis.auth.domain.ExpirationToken;
 import org.openlmis.auth.domain.PasswordResetToken;
 import org.openlmis.auth.domain.User;
 import org.openlmis.auth.dto.referencedata.UserDto;
 import org.openlmis.auth.exception.ValidationMessageException;
 import org.openlmis.auth.i18n.ExposedMessageSource;
+import org.openlmis.auth.repository.EmailVerificationTokenRepository;
+import org.openlmis.auth.repository.ExpirationTokenRepository;
 import org.openlmis.auth.repository.PasswordResetTokenRepository;
 import org.openlmis.auth.repository.UserRepository;
 import org.openlmis.auth.service.notification.NotificationService;
@@ -37,15 +45,12 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.util.StringUtils;
 
-import java.time.ZonedDateTime;
-
 @Service
 public class UserService {
+  public static final long TOKEN_VALIDITY_HOURS = 12;
 
-  public static final long RESET_PASSWORD_TOKEN_VALIDITY_HOURS = 12;
-
-  static final String RESET_PASSWORD_URL =
-      System.getenv("BASE_URL") + "/#!/resetPassword/";
+  static final String RESET_PASSWORD_URL = System.getenv("BASE_URL") + "/#!/resetPassword/";
+  static final String VERIFY_EMAIL_URL = System.getenv("BASE_URL") + "/verifyEmail/";
 
   private static final String MAIL_ADDRESS = System.getenv("MAIL_ADDRESS");
 
@@ -60,6 +65,9 @@ public class UserService {
 
   @Autowired
   private PasswordResetTokenRepository passwordResetTokenRepository;
+
+  @Autowired
+  private EmailVerificationTokenRepository emailVerificationTokenRepository;
 
   @Autowired
   private NotificationService notificationService;
@@ -99,6 +107,10 @@ public class UserService {
     }
     dbUser = userRepository.save(dbUser);
 
+    if (isNewUser && referenceDataUser.hasEmail()) {
+      sendEmailVerificationEmail(dbUser, referenceDataUser.getEmail());
+    }
+
     return dbUser;
   }
 
@@ -109,9 +121,37 @@ public class UserService {
    * @return password reset token
    */
   public PasswordResetToken createPasswordResetToken(User user) {
-    PasswordResetToken token = passwordResetTokenRepository.findOneByUser(user);
+    return createExpirationToken(user, passwordResetTokenRepository, arg -> {
+      PasswordResetToken token = new PasswordResetToken();
+      token.setUser(arg);
+      token.setExpiryDate(ZonedDateTime.now().plusHours(TOKEN_VALIDITY_HOURS));
+
+      return token;
+    });
+  }
+
+  /**
+   * Creates (and deletes existing) email verification token for given user.
+   *
+   * @param user token's user
+   * @return email verification token
+   */
+  private EmailVerificationToken createEmailVerificationToken(User user) {
+    return createExpirationToken(user, emailVerificationTokenRepository, arg -> {
+      EmailVerificationToken token = new EmailVerificationToken();
+      token.setUser(arg);
+      token.setExpiryDate(ZonedDateTime.now().plusHours(TOKEN_VALIDITY_HOURS));
+
+      return token;
+    });
+  }
+
+  private <T extends ExpirationToken> T createExpirationToken(User user,
+      ExpirationTokenRepository<T> repository, Function<User, T> creator) {
+    T token = repository.findOneByUser(user);
+
     if (token != null) {
-      passwordResetTokenRepository.delete(token);
+      repository.delete(token);
       // the JPA provider feels free to reorganize and/or optimize the database writes of the
       // pending changes from the persistent context, in particular the JPA provider does not
       // feel obliged to perform the database writes in the ordering and form implicated by
@@ -119,13 +159,10 @@ public class UserService {
 
       // the flush() flushes the changes to the database so when the flush() is executed after
       // delete(), sql gets executed and the following save will have no problems.
-      passwordResetTokenRepository.flush();
+      repository.flush();
     }
 
-    token = new PasswordResetToken();
-    token.setUser(user);
-    token.setExpiryDate(ZonedDateTime.now().plusHours(RESET_PASSWORD_TOKEN_VALIDITY_HOURS));
-    return passwordResetTokenRepository.save(token);
+    return repository.save(creator.apply(user));
   }
 
   /**
@@ -136,22 +173,44 @@ public class UserService {
    * @param isNewUser whether the user was just created
    */
   public void sendResetPasswordEmail(User user, String email, boolean isNewUser) {
-    UserDto referenceDataUser = userReferenceDataService.findOne(user.getReferenceDataUserId());
     PasswordResetToken token = createPasswordResetToken(user);
-
-    String[] emailBodyMsgArgs = {referenceDataUser.getFirstName(), referenceDataUser.getLastName(),
-        RESET_PASSWORD_URL + token.getId().toString()};
-    String[] emailSubjectMsgArgs = {};
 
     String subjectMessageKey = (isNewUser) ? ACCOUNT_CREATED_EMAIL_SUBJECT
         : PASSWORD_RESET_EMAIL_SUBJECT;
 
+    sendEmail(user, email, token, subjectMessageKey, PASSWORD_RESET_EMAIL_BODY, RESET_PASSWORD_URL);
+  }
+
+  /**
+   * Sends email verification email.
+   *
+   * @param user      the user whose email is being verify
+   * @param email     recipient's email address
+   */
+  private void sendEmailVerificationEmail(User user, String email) {
+    EmailVerificationToken token = createEmailVerificationToken(user);
+    sendEmail(
+        user, email, token,
+        EMAIL_VERIFICATION_EMAIL_SUBJECT, EMAIL_VERIFICATION_EMAIL_BODY,
+        VERIFY_EMAIL_URL
+    );
+  }
+
+  private void sendEmail(User user, String email, ExpirationToken token,
+      String subjectKey, String bodyKey, String bodyUrl) {
+    UserDto referenceDataUser = userReferenceDataService.findOne(user.getReferenceDataUserId());
+
+    String[] bodyMsgArgs = {
+        referenceDataUser.getFirstName(),
+        referenceDataUser.getLastName(),
+        bodyUrl + token.getId().toString()
+    };
+    String[] subjectMsgArgs = {};
+
     notificationService.send(new NotificationRequest(
         MAIL_ADDRESS,
         email,
-        messageSource.getMessage(subjectMessageKey, emailSubjectMsgArgs,
-            LocaleContextHolder.getLocale()),
-        messageSource.getMessage(PASSWORD_RESET_EMAIL_BODY, emailBodyMsgArgs,
-            LocaleContextHolder.getLocale())));
+        messageSource.getMessage(subjectKey, subjectMsgArgs, LocaleContextHolder.getLocale()),
+        messageSource.getMessage(bodyKey, bodyMsgArgs, LocaleContextHolder.getLocale())));
   }
 }