+++ src/main/resources/db/migration/20180607110415124__added_email_verification_tokens.sql
@@ -1,0 +1,10 @@
+CREATE TABLE email_verification_tokens (
+    id uuid NOT NULL,
+    expirydate timestamp with time zone NOT NULL,
+    userid uuid NOT NULL
+);
+
+ALTER TABLE ONLY email_verification_tokens
+    ADD CONSTRAINT email_verification_tokens_pkey PRIMARY KEY (id),
+    ADD CONSTRAINT email_verification_tokens_userid_uk UNIQUE (userid),
+    ADD CONSTRAINT email_verification_tokens_userid_fk FOREIGN KEY (userid) REFERENCES auth_users(id);